releases: ensure rsync timers are removed when switching backends
With the current "unless" code there is nothing that removes
a existing rsync timer from a releases* backend when we switch
from one datacenter to another.

A resource gets created or is ignored but never removed.

Therefore remove the "unless" part and set the ensure parameter
to present or absent depending on this being the primary server
or not.

This is to fix monitoring alerts about failed timers trying to
sync in the wrong direction, which fails because firewall is closed.

Bug: T330960
Change-Id: I15def939a06f7dfb16835ebf2ddce6da2055d16f