Disable all gobblin jobs to allow for HDFS maintenance
We wish to be able to pause ingestion to HDFS by temporarily disabling
all gobblin jobs, on both the production and test Hadoop clusters.

This will first be used in order to support the switch upgrade work
in eqiad.

DO NOT MERGE this until shortly before the maintenance window of
March 7th - 14:00-16:00 UTC.

We will also need a patch to revert it once the maintenance is complete.

Bug: T329073
Change-Id: Ia34876e2a6d0d1bfb114f94dead1dc2086549ece