P:rsyslog: manage /etc/logrotate.d/rsyslog
buster and bullseye have different configuration for the syslog
logrotate file.  but buster rotating /var/log/syslog daily
and bullseye rotating weekly. as such start managing this file so that
the config is the same for all hosts

At the same time change all log files so they are rotated daily.  The
logrotate file can be controlled via hiera for the odd occasion where it
is required to override the default

Hosts: basic
Hosts: O:kerberos::kdc
Change-Id: Ic30d13902c96b7ea42dcd5e67f03f36523f636a1