peopleweb: ensure each user automatically gets a public_html dir
Using admin::unique_users(['all-users']) ensure that each user
on people* machines automatically gets a public_html
directory inside their home directory.

Owned by the user and group "wikidev" and 0755 is the current status, so
just trying to keep it that way. Yes, puppet just adds the executable
bit, so writing 0644 would do the same here.

In the past users had to manually create this and this way it simply
exists and it should make it easier to sync between multiple people*
hosts.

Previously we could have the situation where a user has a public_html
in eqiad but not in codfw or vice versa and then data syncing could
fail.

Bug: T330091
Change-Id: I902b84c8b5a86cee9d54e3db10073aa255a13003