logstash: Stop apache2-htcacheclean.service via Puppet
apache2-htcacheclean is a service which gets shipped by the apache2
package. It's not enabled by default, but only enabled in the postinst
if the selection of Apache mods needs it. While this is the case for
the OpenSearch API proxy, the service isn't actually needed, so make
sure it's stopped

Hosts: O:logging::opensearch::collector
Change-Id: Ic7b85e63dffd02e3454809492f5e1be7c9e3fa57