arclamp: Enable profile::auto_restarts::service for apache2-htcacheclean
apache2-htcacheclean is a service which gets installed by the apache2
Debian package, but the service is only enabled/started if one of the
enabled Apache modules needs it. This is the case for arclamp.

Automatically restart it if underlying runtime library dependencies
have been upgraded. The restart is splayed across each day (on Mo-Fr).

Tested on arclamp1001 following the latest apr-util security update.

Bug: T135991
Hosts: arclamp2001.codfw.wmnet
Change-Id: I7cd306e243aa1604c495e0ff6a1f8a353251e47d