IDM: Enable profile::auto_restarts::service for apache2-htcacheclean
apache2-htcacheclean is a service which gets installed by the apache2
Debian package, but the service is only enabled/started if one of the
enabled Apache modules needs it. This is the case for Bitu.

Automatically restart it if underlying runtime library dependencies
have been upgraded. The restart is splayed across each day (on Mo-Fr).

Bug: T135991
Change-Id: Ie8a488980b9cf1513a8825ed565b2dc254a76817