Add the aphlict config on aphlict2001.codfw
Currently the aphlict and phabricator config is hand-deployed on
aphlict1001.eqiad. We want to manage this with puppet. To do
this we will apply a puppetised config on aphlict2001, which is
currently not in use.

Once we have this deployed, and ensure that we can deploy phabricator to
these hosts with scap, we will work on retro-fitting the puppet config
onto aphlict1001. Eventually, the goal will be to reimage these hosts
and ensure they come up correctly with puppet from boot.

Extra changes in this are moving some script deployment out of the
phabricator main and into the config module to satisfy dependencies.
Perhaps config.pp should be renamed to common, or scripts should be
broken out to their own module. Feedback welcome!

Bug: T322369
Change-Id: I9fa2d06e827e5043e840aab5d1cbbc9aec2aa923