smart: override unit to make systemd wait longer
Smartd takes a while to start up, particularly on nodes with a lot of
disks (like the swift backends, particularly the new ones where the
disks are all JBOD) - and the systemd default timeout of 90s is too
short, leaving to systemd killing smartd and marking the unit as
failed.

Increase the timeout to 300s / 5 minutes.

Hosts: ms-be2070.codfw.wmnet
Bug: T326352
Change-Id: I73688616f632f502318e465fe699b4a0ad2792fa