P:blackbox_exporter: update client auth checks to use local certs
Currently the `prometheus::blackbox::check::{http,tcp}` checks use
$facts['puppet_config']['hostcert'] and
$facts['puppet_config']['hostprivkey'] to get the values for the client
auth certificate and key.  however theses values are evaluated as the
resource is exported and as such point to the certificate files of the
server been monitored.  however as the check is executed on the
monitoring server we need this value to point to certificates that exist
there.

To resolve this issue i have added puppet::expose_agent_certs to the
prometheus servers so that we have a deterministic path name that we
can use as defaults for the client auth key and cert.

AFAIK no checks currently use theses parameters so this should be mostly
a no op, other then the extra /etc/prometheus/ssl folder on
prometheus servers.

Hosts: P:prometheus::blackbox_exporter
Change-Id: I30722bba752a3e091823567bd938211c43168fe5