striker: Bump container version to 2023-03-10-212005-production
Follow up to 4869a37 and 2ab962f. The Striker build in the container
includes config to use the new 'member' and 'reader' as the expected Keystone
roles for Toolforge admins and tool maintainers respectively. It also
includes the fix to striker.openstack.Client.users_by_role that removed
a bad assumption that each user would only ever have a single role.

Bug: T330759
Bug: T331674
Change-Id: I972f73e2207ec54f1d39f7edade27fae4adfcf84