install_server: use second pair of disks for /srv/gitlab-backup
The future GitLab hosts will have 4 disks (2x 900GB and 2x 1800GB).
This change updates the partman config. The first pair of disks is still used
for the root file system and a registry volume. The size of the root file
system is increased slightly due to more available space.
The second pair of disks is used entirly for storing/creating backups at
/srv/gitlab-backup. For this disk no lvm is configured (mostly due to lack
of knowledge how to configure two different lvm layouts). But it's also
not really needed.

Bug: T330172
Change-Id: I0724186d9372b4746153eb41e7e8fbc5a45d5d87