netboot: fix a few syntax errors
Remove,
  10.64.36.1) echo subnets/labs-hosts1-c-eqiad.cfg ;;
which is the wrong vlan and masks,
  10.64.36.1) echo subnets/analytics1-c-eqiad.cfg ;;

Fix the globbing for druid1009 - druid1011, as ranges can only include a
single character.

Also replace the leading tab character for the druid* rule with spaces

Remove mwdebug[0-9]* which is masked by mwdebug* with the same config.

Fix a missing square bracket on sretest100[1-2]

Bug: T330495
Hosts: O:installserver
Change-Id: I26c49170898ce7b40a201d7fe37ed3f69433057d