exim: remove wikimedia.com from wikimedia_domains
Per ticket comment and the change that added this back in
2014 (I1d220ee9895bfcec1) email for this domain used
to be handled by OTRS (now called VRTS).

That is the system behind https://ticket.wikimedia.org where
volunteers handle support requests for various mail queues.

Years later in 2021, the domain wikimedia.com was then removed
from Wikimedia's DNS servers in c33805a6b92a624a0d7 with the
comment that this was migrated to Enterprise's DNS servers.

This was handled in ticket T281428 back then.

Now the MX record for wikimedia.com points to AWS (see 'dig MX
wikimedia.com') and WMF DNS servers don't know about this anymore.

But, the domain was not removed yet from this file that lists
"wikimedia_domains", which then influeces how our exim servers
are configured.

Now in 2023 a new ticket T331676 was opened because mails from
Enterprise mail servers don't reach wikimedia.org servers.

As Jesse Hathaway points out while debugging for that, sender
verification fails with Address wme_support@wikimedia.com does not
exist.

The expectation is that removing the domain from this file turns
off the sender verification and since this is not supposed to be a
"wikimedia domain" (as in the non-profit org) anymore it should
also be the right fix anyways.. and probably should have already
happened when the domain was removed from our DNS servers in 2021.

Bug: T331676
Bug: T281428
Change-Id: I4ce6e5c8985fb97f4f118fb295eedac0a4ae8204