Adjust OSPF Icinga check to ignore OSPFv3 if zero ints configured
This is perhaps a naive approach here, and we may want to instead
add a command-line flag to enable checking for OSPFv3 instead.

But it works well enough, and I can't forsee serious issues.  A
situation where we inadvertently configure ZERO interfaces for
OSPFv3, where we should have them (say on a core router), seems
unlikely to happen.

But it's no problem to make it a flag instead if we think that's
best.

Bug: T315053
Change-Id: Ib284bef9111793cc3d3e563563f13a14bc870b42