cloudcontrol2001-dev: introduce cloudlb support
This patch migrates cloudcontrol2001-dev to work with the cloudlb setup.

It means there will be an external HAproxy, and they should communicate using the
cloud-private vlan subnet.

I've tried to figure out which hiera values makes sense with the cloud-private
FQDN and which one doesn't. See patch comments for details.

Bug: T332153
Bug: T336236
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I37c9a8717fbebc726e50d1c72c441c7b7306dd38