trafficserver: make routing to mw on k8s more manageable
The use of regexes to match domains to route to
mediawiki on k8s was unsustainable already having migrated just two
wikis. Thus we add a new lua module where we can both add new domains
easily and add a different logic if needed.

Please note: we haven't added the ability to reload the configuration
from a config file for now to make the first implementation as simple as
possible. We might move to use it at a later time if we see the rate of
change is too high.

For now I feel like we don't have a good enough model of how we want to
transition the largest part of the traffic to allow us to define a config
format properly.

Bug: T331318
Change-Id: Ifde6e1881f31397986d88e781a0f9964afe359db