Omit Python 3.7 in the analytics cluster after bullseye
We have been installing Python version 3.7 as a common package on the
analytics cluster through stretch and buster. This version of pythin is
not available on bullseye and we do not plan to forward-port it.

Therefore this commit skips the requirement to add the python37 profile
for bullseye hosts or later.

Hosts: auto
Bug: T329363
Change-Id: I2c7f3ef50a8eff25c04263003c6daa8634f68363