multi-dc: Use primary for OAuth for both URL forms
Using the /authorize OAuth endpoint with a non-pretty URL is
fragile (T74186) and not recommended, but many applications do
it nevertheless so make sure it's sent to the primary DC.

In theory, the /initiate and /token endpoints do not work at all
with a pretty URL (T59500), but it's hard to be sure so it's less
hassle to just support it anyway.

Follows up Id8799bae90590f082845c3e1c966d9fa25b96b47.

Bug: T313578
Bug: T332650
Change-Id: I587a6e2f41f552b1a8234ccb781a686e9bf8fd3c