Allow hive on bullseye to install and use the correct packages
On bullseye we need to allow the installation of the python2.7 minimal
packages when hive and hive-hcatalog packages are installed. On the
analytics_test cluster this includes the workers, the coordinator, and
the client roles.

However, we also want to install the python-is-python3 packages in order
that it uses the correct version of python at runime.

Hosts: auto
Bug: T329363
Change-Id: I38f86a29abb86c990e9f3d204114aaaef3cd7902