zuul: fix up service enable and ensure
This align the Zuul merger and server service parameters to be
consistent and to match the systemd::service service_param.

Zuul server
===========

`service_ensure` accepted a Boolean value and had a default set to
`running`. Changed it to Stdlib::Ensure::Service to drop Boolean from
the type since that was never used.
Remove the `running` default from the profile requiring it to be
explicitly defined (that is done from the role currently).

`service_enable` duplicated the `Wmflib::Enable_Service` type.

Zuul merger
===========

This had a single `ensure_service`, align it with the Zuul server and
switch to `service_ensure` and `service_enable`.

Those changes let us pass the hiera setting directly to systemd::service
service_param and thus drop some conditional statements and intermediate
"real" variables.

Alphabetically sort `enable` and `ensure` parameters in the Puppet
classes as well as in the Hiera yaml files.

The sole change is the `zuul::merger` service is now explicitly enabled.

Bug: T324659
Hosts: contint2001.wikimedia.org
Hosts: contint1002.wikimedia.org
Change-Id: I80db69c52bbc4e0d152debde2309851e16b7498a