Use the spark3 shuffle jars to yarn on a test host
To-date we have been using the Spark2 network shuffle service jars with
YARN, even after the update to Spark3 in production.

This commit starts to test the use of the spark3 shuffler jars on a host
in the test cluster, whilst retaining the use of the spark2 shuffler
everywhere else.

Hosts: an-test-worker1001.eqiad.wmnet,an-worker1078.eqiad.wmnet
Bug: T329363
Bug: T295072
Change-Id: I03842be8d6914a2664f7184e563c2b409ba9eac8