dbbackups: Setup db1145 as a backup source replacement for db1150
db1145 used to contain s4 and s3, but was passive. Remove s3
data and deploy s5 there so it can take over db1150 backup source
role after the latter crashed.

Set both dumps and snapshots for both s4 and s5 to point to db1145
(A data transference will have to happen for s5).

Bug: T332708
Change-Id: I63b73da9dae152b18a3151cd263aa8025d41d09c