buildkitd: Isolate build container user/process/network namespaces
Refactor buildkitd to run in privileged mode which is required for
buildkitd to create separate user/process namespaces for build
containers, and enable CNI to create a pool of network namespaces.

The former approach to running in rootless mode afforded privilege
separate between buildkitd itself and the host system, but effectively
afforded no isolation between build container processes. Given that the
latter are the most untrusted processes of our system, and not buildkitd
itself, we should solve for the best isolation possible between them.

Note that dockerd already runs buildkitd in a privileged mode internally
when building images (via `docker build` or `docker buildx`), so running
our own buildkitd in privileged mode is comparable to that. However we
are now giving runner hosts remote access to a process with higher
privileges on the host (not nearly the privileges of dockerd, but still
higher than before), so we need to stay on top of security updates to
buildkitd as we do with dockerd.

Bug: T332804
Change-Id: Ic5cf6fa7c15693c79e4e94dd45ad291d6194960a