mediawiki::errorpage: rationalize usage
We had a couple places where we wanted just the content of the errorpage,
not to write a file to disk. As a consequence, we were calling the template
directly in other modules.

To fix this, a function is created that returns the string value, and also
can allow us to do some degree of parameter checks using types.

Change-Id: I256596d655af125dca6cfc91d7ff8e17cdbdd954