apache2: Use systemd provider
Prior to this commit we used the debian provider for apache2, which was
the sysvinit system prior to systemd.

After this change we use the systemd provider on bookworm and newer
hosts.

I'm not sure why the apache2 service was set to debian, since it should
have been the default.

Bug: T331706
Hosts: O:mediawiki::canary_appserver, O:mediawiki::appserver, O:lists
Change-Id: Ie56678c95eccacc3b0255926a2ba7a58459dabf4