etherpad: replace Icinga with Prometheus monitoring
The goal is to remove all custom Icinga checks for our services
and also make sure each service is creating tasks if triggered.

In this case we already have one prometheus check but now
we want to also replace the other.

They both check etherpad.wikimedia.org but on different ports.

On port 7443 we have envoy, which terminates TLS and listens on IPv4.

On port 9001 we have nodejs, behind envoy, which speaks http and listens
on IPv6.

So use the server_name parameter so that the title can be something
different and account for the differences in protocols.

And generally one can argue if we really want to monitor both envoy
and backend..but theoretically they can fail separately,so why not.

Bug: T329587
Bug: T331901
Change-Id: I42f6c7376080a331265555834fc3eba8242f0327