releases-jenkins: replace Icinga with Prometheus monitoring
The goal is to remove all custom Icinga checks for our services
and also make sure each service is creating tasks if triggered.

The higher goal is to help Observability to get off of Icinga.

What this adds:

- automatic task creation
- checks for a content string, not just status code

What this removes:

- IRC notification
- the notes_url / link to Wikitech runbook

If there is a feeling we should keep IRC notifications we can
toggle that by creating another severity level.

Probably makes sense to agree on this for all or none of our checks
and previously we said "task only" is our new way to go, at
least for the collab team.

I wonder how releng feels about that though for some of the services.

I also miss the notes_url parameter since we have put effort in
the past into "every check should have a runbook URL", so asked
observability about this on IRC.

Bug: T329587
Bug: T331901
Change-Id: I9a6b093efbe406c7a1d76b570e5465d73172a4ed