peopleweb: replace Icinga with Prometheus monitoring
The goal is to remove all custom Icinga checks for our services
and also make sure each service is creating tasks if triggered.

The higher goal is to help Observability to get off of Icinga.

What this adds:

- automatic task creation
- checks for a content string, not just status code

What this removes:

- IRC notification
- the notes_url / link to Wikitech runbook
- cert expiry monitoring but that should be a concern at acme_chief
  level nowadays, we should not need the separate check anymore here

We are checking TLS against envoy which terminates TLS on the host,
not the apache behind it. And envoy only listens on IPv4 as of today.

The TODO to do this only once is outdated. Nowadays we actually want
to check also passive instances. So same virtual host checked against
all backends is desired.

Bug: T329587
Bug: T331901
Bug: T316022
Change-Id: I87496d00ef7b893779fdd79896e5310c0388fd5a