alertmanager: send sre-collab alerts to -operations and -sre-collab
We already have IRC alerts for alerts with severity "task" but they
were still sent to a test channel.

Use the actually new channel now and try to send to -operations as well.

That way all existing Icinga alerts that used to alert in -operations
will still alert in -operations when replaced with Prometheus
monitoring.

Bug: T329587
Bug: T331901
Bug: T316022
Change-Id: I239e8981ffefd5dc9cbaf54af6d26913753ec8a8