admin: add Barakat Ajadi as ldap_only_admin (wmf group)
To be added to the 'wmf' LDAP group for Grafana access.

More reasoning per ticket.

An intern with a -ctr email address, therefore expiry contact
and expiry date are added.

Lariss said "until July" so for now assuming last day of June
is expiry date.

To be added to wmf group in LDAP after/with merge.

Bug: T332868
Change-Id: I29c93cef31015886aebf6461b0e43a8637d8227d