alertmanager: delete unused serviceops-collab receivers
We don't use these receivers anymore. Currently all we
are using is the "task" severity which does ticket creation
and IRC notifications.

We are going to add another type that includes mail to releng
and or other things in the future but it won't be these
"critical" vs "warning" ones that are using a test IRC channel.

Bug: T329587
Change-Id: I607f179ec312ea66fcc526ef6a5fc1b7725c1924