noc: replace Icinga with Prometheus monitoring
My guess is there is nowadays no more point in monitoring
the expiry of TLS certificates specifically on noc.wikimedia.org.

I noticed there does not seem to be a specific team/receiver
in alertmanager for serviceops team but there is this one
literally called "noc-ircmail" which means IRC in -operations
and email to noc@.

So it's either that or "sre-irc" or "sre-page" or something new.

What level of alerting should it have if https://noc.wikimedia.org
is down?

The general goal here is just to help with converting alerts
and I ran across this, unlike other services it's not considered
owned by sre-collab.

Bug: T331901
Change-Id: I3ec118f4467943050c8e58acf92c7b24fea55f01