docker_registry_ha: Do not lose original request URI during JWT auth
Duplicate the `set $auth_request_path` directive in the nested location
that matches exact blob/manifest digests to ensure JWT authorizer has
the original request path.

Nginx is quite strange in how it handles nested locations and which
directives from parent `location` blocks apply and which don't, and
there is nothing in the documentation for individual directives that
specifies when that is the case.

In this case, for example, consider this pared down part of the former
configuration.

    location ~ ^/v2/(.*) {
      set $auth_request_path $1;

      # ...

      auth_request /auth;
      auth_request_set $auth_status $upstream_status;

      proxy_pass http://registry;
      # ... (proxy options)

      location ~ ^/v2/(.*)/(blobs|manifests)/sha256:(.*)) {
        proxy_pass http://registry;
        include /etc/nginx/registry-nginx-cache.conf;
      }

One might assume that the `set $auth_request_path $1` at the top of the
parent location would apply to the nested location, since the other
directives like `auth_request` seem to (requests to exact digests are
being auth'd). However, the `set` does _not_ apply. Locally reproducing
this configuration and setting `error_log` to `debug` confirms this.

    2023/03/29 15:41:45 [warn] 27#27: *3 using uninitialized "auth_request_path" variable ...

This change employs a short term workaround which is to duplicate the
`set` line in the nested location as well. A longer term solution might
be to refactor the nested location configuration as server-level
locations to make behavior more predictable.

Bug: T322453
Change-Id: I02eb514c00a4a54dcfceac396cb3c1d25710fbf6