ci::master: add parameter to enable/disable monitoring of jenkins/httpd
We are monitoring if https://integration.wikimedia.org is up on all
CI master (contint*).

On the new host contint2002 the production role has just been added
but jenkins is still masked.

This means production alerts on IRC and tickets in Phabricator now.

Add a parameter to turn it off based on the hostname and make the
default that it's enabled.

Just like other existing things that are disabled here for now.

Bug: T324659
Change-Id: I36be63bed379a35f68d8d09f1992300e8e639eca