microsites: do not use TLS when monitoring commons-query.wikimedia.org
This site has a weird setup where it goes to Apache cluster
for a rewrite but ALSO exists as a micro site on miscweb*.

Except the host name was never added to the TLS cert for hosts
on miscweb!

So this is the only thing that exists as a virtual host but
you will get a certificate error when trying to access it.

This is not a great situation given how much effort we put into
terminating TLS at the backends with envoy for everything else.

But for now I just need to fix monitoring alerts.

follow-up to I0dfe7791021e257a7

Bug: T327976
Bug: T333510
Bug: T333507
Change-Id: Ic2d012fe429ce641f5dd61784d2d422ad966ff82