pybal: port check_pybal_ipvs_diff.py to urllib2
check_pybal_ipvs_diff.py uses the requests library for the HTTP request
part in the script.  This is fine in buster where python-requests
(Python 2) is part of Debian. However, the python-requests package does
not exist in bullseye and we did not backport the package for Python 2
when the other PyBal migrations took place (see [0]).

So that leaves us with two options for check_pybal_ipvs_diff:

1. Port the script to Python 3: While we can port the rest of the script
   and simply use python3-requests instead, we also depend on
   python-prometheus-client (Python 2) and hence the port to Python 3 is
   even more involved unless we are willing to update PyBal's code (so
   that we can use python3-prometheus-client; python-prometheus-client
   was built for Python 2 in [0]).

2. Build the python3-requests package for Python 2, adding to the list
   of packages built for PyBal in bullseye. Since this package won't be
   used elsewhere, this is also not worth pursuing.

This commit therefore simply removes the requests package dependency and
uses urllib2 instead. This also helps us proceed without making extra
changes to maintain the compatibility between buster and bullseye LVSes.

[0] - https://phabricator.wikimedia.org/T321309#8717328

Bug: T321309
Hosts: lvs4009.ulsfo.wmnet
Change-Id: I369ca80c3237977d14de2228eec2c3314a14421a