mediabackups: Add static console port for easier remote management
MinIO Console management port used to be shared with the regular
s3 api (9000). Since some versions, this is handled on a separate
port, dynamically allocated, and for backwards compatibility,
it is redirected from the original port.

However, to make easier accessing through port redirection (where
only one port is forwarded), it is better to define statically
a fixed port for it: 9001.

Also open the firewall to it from workers hosts.

Bug: T306602
Change-Id: Ib97ed0b2e9eccb025a03593c0e48ec5cd00d5eb6