pybal: don't install python3-requests on bullseye hosts
With the rewrite of check_pybal_ipvs_diff.py to not use the requests
module, remove the installation of python3-requests on bullseye.

Even though the package will also not be required for buster, keep it
installed so as to not potentially break stuff. All LVS hosts are to be
upgraded to bullseye anyway so it should be fine.

Bug: T321309
Hosts: lvs4009.ulsfo.wmnet
Change-Id: I836b6f565ec1d9a32ea53b8a9d64512d31fedbc0