modules::profile::manifests::airflow.pp: add plugins_folder path
Until now, the plugins_folder was configured by default to point
to {AIRFLOW_HOME}/plugins. That was fine, because we did not
have any custom plugin implemented.

Now, though, we created our first plugin, and we'd like to place
it in a location where all Airflow instances can make use of it.
So, the Airflow config should point the plugins_folder property
to <repository root>/wmf_airflow_common/plugins instead.

Bug: T324485
Change-Id: Icca7d022b6412599a7a18edf84cbf0404ec33d48