jupyterhub-conda: Fix incompatibility with HDFS-FUSE mount
On bullseye systems the jupyterhub-conda service fails to start with a
status=226/NAMESPACE error. This happens when it tries to remount the
whole host's filesystem as read-only.

The InaccessiblePaths option can be used to exclude /mnt/hdfs from the
systemd unit when it starts, avoiding the error condition.

However, when users try to spawn their own jupyterhub-singleuser
service this step would also fail due to the way that the systemdspawner
attempts to create a read-only permissions overlay on the root file
system, with named exlusions for read-write directories.

See https://github.com/jupyterhub/systemdspawner#readonly_paths for more
documentation on this feature.

We have to disable this feature in order for the singleuser services to
start, which means that users will be relying on the native file system
permissions of the host, as opposed to the overlay.

The /mnt/hdfs volume is still available to user of jupyterhub on
bullseye with these changes.

Bug: T333511
Change-Id: I5d17efe560f0acdc342161273670e447979f56b7