Switch kafka-main1001 broker's TLS cert to PKI
This is the first broker to use PKI for kafka main. The Kafka
clients can fallback to another broker in the list in case
the TLS handshake fails, so even if some clients are still using
the Puppet CA certificate as CA bundle they shouldn't stop
working. We'd have the possibility to monitor (via tcpdump/tshark)
those handshake failures and follow up as needed.

Bug: T319372
Change-Id: I424b421e3b3ee04e4994d1741cb72221aae12a15