etherpad: remove process monitoring
I think we don't need this check. We are already monitoring
if https://etherpad.wikimedia.org is up.

I see little value in having another check for the process
running. Either it's up or it's not.

And if it's down it would just create duplicats alerts proabbly.

This way we don't have to worry about how to replace it with a
Prometheus check to get off of Icinga.

Bug: T331901
Change-Id: I6382727193bea60acb96c50a13dfcf4788edaf32