profile::hue: restart envoy only when CAS is enabled
Envoy is added only in the non-test-cluster roles, but its
auto-restart unit is always added in profile::hue. To avoid
unnecessary alerts, this patch enables it only when CAS is
enabled. For the moment it seems a quick stopgap, maybe a more
in depth refactoring will be needed in the future.

Change-Id: I5d7348b73b93e50177867f8644c242d70f1102d3