trafficserver/wdqs: switch query-preview.wikidata.org to new backend
I would like to switch this service from an older to a newer backend.

This is basically a distro upgrade of the backend. Don't get confused
by the naming.  "app" or "sites" are just labels here, the point
is to move from miscweb1002/2002 (buster) to miscweb1003/2003 (bullseye).

Most other sites have already moved.

Since this is just a static site and an apache there is no expected
problem here, also not with the TLS certs.

The one thing that makes this special to me as the one switching
sites over is that it requires edits to files other than ATS
backend.yaml.

So for the WDQS part I would like a review that also serves as a
heads-up.

I noticed there are "test", "internal" and "public" and assume
test and internal are query-preview while public is query.

I want to do the same thing to query.wikidata.org after this is
done and we confirmed no problems.

We have "bttpbb" tests on deployment servers that check both a
query-preview and a query.wikidata URL and they work on old and new
backends.

The virtual host and document root content are in place via puppet
and git pull run by puppet.

Bug: T331896
Bug: T330090
Bug: T333656
Change-Id: I876e0b1cb1f3bae15b479287fae86b47ab6d54b2