mediawiki::scap: Ensure Exec['fetch_mediawiki'] resource always exists
Ibf105893e8f7ed7972d217299258a19399853860 made it so that if
profile::mediawiki::scap_client::deployment_dir_linked_to_staging_dir
is true, the Exec['fetch_mediawiki'] resource would not exist (because
the scap pull operation isn't wanted in that case), but this breaks
class scap::master which has a 'before' reference to the
Exec['fetch_mediawiki'] resource.

This commit retains the Exec['fetch_mediawiki'] resource but ensures
that it runs a no-op command when
profile::mediawiki::scap_client::deployment_dir_linked_to_staging_dir
is true (and running on a scap master).

Bug: T329857
Change-Id: I84b11bdb95a9b229d31c70dff51a2de0bf03684c