admin: hashar: some more git aliases
`git pending` to easily find what is waiting in the upstream repository,
quite helpful when doing a deployment to verify what got merged in
Gerrit since the current state on the deploy server.

`git amend` as a shortcut to `git commit --amend`, also allow empty
commits which are sometime helpful.

Change-Id: I5714c8db37b6324a5cc6b31eacd805b583e1a1a9