P:toolforge::k8s::etcd: load list of control nodes from PuppetDB
Instead of relying on a static list of nodes in hiera, make the etcd
profile query that from PuppetDB instead. This means that the only
etcd-related thing the new 'add control node' cookbook has to do is to
run Puppet on all of the etcd nodes.

Unlike the HAProxy profile, where we want to be able to manually
control which nodes are receiving traffic, here using PuppetDB works
great since it includes nodes that are still being set up and nodes that
are being removed, as both of those do still need etcd connectivity.

Bug: T274499
Change-Id: If873d6c8f90a7a3979b6209a735b9974cead1983