profile::kafka::{broker,mirror}: refactor TLS settings
All clusters are using PKI now, so we can refactor the broker's profile
to set PKI by default and remove migration settings.

Most notably this change will modify the super.users listed in
the kafka config, since we don't need anymore to add the old CN
contained in the Puppet TLS certs.

Set also PKI TLS certs generation to false for cloud, so that
every project is free to enable it if needed.

Change-Id: I355943861f111d1fe7538bc202252e93b7e360cc