install_server: simplify gitlab disk layout, drop lvm, use four SSDs
This change tries to use all four disks on the GitLab hosts and create
two RAID 1:

/	:		ext4, RAID1, 900GB on sda & sdb
/srv/gitlab-backup :	ext4, RAID1, 1800GB on sdc & sdd

To achive this, the lvm config was removed and both RAIDs have one
single partition only. The dedicated partition for /srv/registry is
nice to have but makes the setup more complex. Furthermore the
registry is not enabled (T304845).

Bug: T330172
Bug: T333674
Change-Id: I2c19491597e1b0a6f11731ddc97adfc25a6aff4f