Revert "mediawiki::scap: force creation of the symlink when enabled"
This reverts commit f16434758cd8d42a7e90daf0bc980ac610ad14a1.

With force => true, if the directory exists puppet will back up every
file in it before deleting it.  This takes an eternity for
/srv/mediawiki.  It is better to not force the operation (which leaves
the directory in place if it exists) and to have a privileged user run
"mv /srv/mediawiki /srv/mediawiki.old" on deploy servers as needed as
T329857 is rolled out.

Bug: T329857
Change-Id: Ic55cb8ebf0b883b8e1fd55b287b2254f88230637