zuul: disable monitoring for disabled merger service
contint2002 is being provisionned with all the application services
disabled. The monitoring for the zuul-merger process is alarming since
we never had the use case to disable that service before.

Disable monitoring when the service is not enabled, this align the Zuul
merger Puppet classes with the Zuul server ones.

Hosts: contint2002.wikimedia.org
Hosts: contint2001.wikimedia.org
Bug: T324659
Change-Id: I397c41a9b942893b5468b6be47ae099d2fed7d9a