aptrepo: import AMD ROCm 5.4 to bullseye-wikimedia
New round of upgrades for the ROCm drivers, mostly to keep
up-to-date the software that enables our GPUs. Since the stat
boxes are being reimaged to bullseye, this change avoids to
import the packages to buster. There are GPUs on the k8s DSE
cluster that we want to test, and the nodes are all bullseye.

Note: http://repo.radeon.com/rocm/apt/5.4/dists/ shows various
options, I chose 'focal' since it seems the closest one to
bullseye.

Bug: T295661
Change-Id: I3401a7126f9db90d33c8a8961625db9fe9fec20a