utils: rm hiera_lookup (replaced by puppet lookup)
The hiera_lookup utility has been broken for quite a bit.

At first it refers to a no more existing hiera.yaml file, the config
files have been renamed by 20777c245cff in September 2022:

- ./modules/puppetmaster/files/production.hiera.yaml
+ ./modules/puppetmaster/files/hiera/production.yaml

Then if I adjust the ruby code to point to the new file, it hits another
issue:

  $ ./utils/hiera_lookup --fqdn=contint1002.wikimedia.org \
     --site=eqiad profile::ci::jenkins::prefix
  ...
  bundler: failed to load command: ./utils/hiera_lookup.rb (./utils/hiera_lookup.rb)
  Traceback (most recent call last):
    <snip>
  hiera-3.12.0/lib/hiera/config.rb:32:in `load':
  v5 hiera.yaml is only to be used inside an environment or a module and
  cannot be given to the global hiera (RuntimeError)

Given the utility has been broken for a few months and the functionality
is now built-in Puppet (`puppet lookup`), remove the utility.

The only drawback is people not having access to the Puppet master can
not do Puppet lookup, then the utility has been broken for a few months
already indicating that close to nobody relies on it.

Change-Id: Ie9fc387ea46d24e1f63cba44e8452ca05ddb5909