site: add role(gerrit::migration) to gerrit1003 and fix code
This adds firewalling and rsyncd config to allow copying
data to a new gerrit host before applying the complete
gerrit production role.

Add gerrit1003 to list of allowed destination hosts.

We have used this in the past already and now comes in handy
that we still kept it for next time.

A bug has been introduced into it at some point
where "user_name" is there but not "daemon_user"
and it could not lookup the daemon_user from the main gerrit
role.

Both are the same thing though, 'gerrit2'. Fixing that too,
by renaming parameter and also moving the Hiera values
from role to profile level so they can be found on all hosts.

This is because the migration profile is always included on
production hosts as well. It is needed on all hosts.

Bug: T326368
Change-Id: I28c32fe2f83924c09bec10536d1f82ba727dfa14