vrts: do not use /srv/sqldata as mariadb datadir (cloud, devtools)
When puppet installs mariadb, via profile::mariadb::generic_server,
the default datadir is set to /srv/sqldata.

But first the package is installed, which by default uses /var/lib/mysql,
then a /etc/my.cnf is written with the different datadir /srv/sqldata
and no service restart is performed.

This makes it look like everything is ok.. until for some reason the
service or VM gets restarted. Then suddenly the new data dir is applued
and for the user the data they setup seems to disappear.

This fixes it by setting the default datadir to the same that the package
uses, but only for VRTS, without influencing other users of the generic
mariadb class.

Bug: T329571
Change-Id: I04141c8965b76e8fa572c262a889caf7094c40db