gerrit: remove leftover absent http config
The Apache Proxy for Gerrit used to listen to specific IP addresses
incluiding an IPv6 address. On boot, the IPv6 address is not immediately
available due to Duplicate Address Detection mechanism, but systemd
nonetheless attempts to start Apache which fails to Listen.

The workaround (short of fixing up systemd itself not waiting for IPv6
to be ready) has been to change Apache to listen on any address and add
iptables rules to firewall out on the IPs we don't want to receive any
traffic to.

I'd thus added a Puppet stanza to purge the http configuration snippet,
and given it has now long been removed, we no more need to explicitly
remove it.

Bug: T326125
Change-Id: If9204796c00d49b5355652a359e7330f2b2106ee