gerrit: remove duplicate $gerrit_site definition
The Gerrit site is the base directory for the application. It is defined
in the gerrit class and forged from the daemon user (gerrit2).

The gerrit::proxy profile redefines it for linking images into the Apache
DocumentRoot. To avoid a redefinition, and since a host having a
gerrit::proxy also has Gerrit, reuse the existing variable.

The reduced duplication will slightly simplify relocating Gerrit site.

Bug: T333143
Hosts: P:gerrit::proxy
Change-Id: I8d043e84f746dcad294306b90dcf5e05702b501b